+++ tests/tinytest.R
@@ -1,0 +1,2 @@
+if (require(tinytest, quietly = TRUE)) test_package("ChemoSpec", verbose = TRUE)
+